Added hooks for running code at the start and/or end of a transaction
